requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
reportlab<4.1
python-multipart
jinja2
